pyproject.toml
[project]
name = "lvlm-interpret-medgemma"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "pandas-stubs==2.3.0.250703",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.36.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "scikit-image>=0.21.0",
    "tqdm>=4.65.0",
    "torchxrayvision>=1.2.0",
    "matplotlib>=3.7.0",
    "ipython>=8.0.0",
    "psycopg2-binary>=2.9.0",
    "gdown>=5.2.0",
    "notebook>=7.4.4",
]
